Set override-redirect on the overlay
This makes the overlay window behave correctly with window managers like
i3 that are aggressive about taking control of windows, even when the
"softer" flags like "make this on top", or "make this not resizeable"
tell them not to.

The override-redirect flag straight up disables the mechanism most WMs
use to affect windows, for this window only.

The override-redirect flag does nothing if our window has already been
captured by the WM, so we hide it (which unmaps it, so the WM lets go),
set all our "can't touch this" flags, then re-show it.  It's hacky, but
GTK segfaults if we try to set anything before
`gtk_widget_show_all(window)`.   But it works.