Delete unnecessary draw code
Umm I'm not sure why I would go into the trouble of rendering a
completely transparent rectangle.  Was that necessary at some point to
actually get a GTK to actually make the window transparent?

I don't know, but it's apparently no longer necessary (if it ever was)
because hudkit works without it.