Fix positioning when started on secondary monitor
Previously, we were treating the desktop as an uninterrupted uniform
plane, and optimistically assuming that GTK will always stick our window
at absolute coordinates 0,0 when we create a window.

This assumption does not match reality: by default, GTK positions new
windows in the top-left corner of the currently active physical monitor.
This caused the overlay window to be misplaced when hudkit was started
on a monitor other than the top-left-most one.

Fixes #2