Grab keyboard focus when clickable area hovered
Finally figured out how to do this.  This change makes it possible for
the first time for the overlay window to react to keyboard events, and
so it's possible to make any arbitrary interactive web UI with it!

It's pretty neato.

This is a proof of concept!  There are many minor bugs, like
drag-and-drop completely breaks when trying to drag out of the overlay's
clickable area, and enter/leave events jitter when sweeping the mouse
along edges of the screen.  That kind of thing.

Related to #3 and #5, but not global; no events are duplicated here.